refactor split tools into modules
