+++ tools/mcp.py
@@ -1,0 +1,3 @@
+from fastmcp import FastMCP
+
+mcp = FastMCP("Web Scraper MCP 🚀")
